+++ test/EndToEndSpec.hs
@@ -463,6 +463,22 @@ spec = around_ (inTempDirectoryNamed "foo") $ do
             Paths_foo
         |]) {packageCabalVersion = ">= 1.22"}
 
+      it "includes all generated modules" $ do
+        [i|
+        library:
+          generated-exposed-modules: ABC
+          generated-other-modules: XYZ
+        |] `shouldRenderTo` (library [i|
+        exposed-modules:
+            ABC
+        other-modules:
+            Paths_foo
+            XYZ
+        autogen-modules:
+            XYZ
+            ABC
+        |]) {packageCabalVersion = ">= 2.0"}
+
       context "when inferring modules" $ do
         context "with exposed-modules" $ do
           it "infers other-modules" $ do
@@ -482,6 +498,29 @@ spec = around_ (inTempDirectoryNamed "foo") $ do
                 Paths_foo
             |]
 
+          it "doesn't doubly include generated modules under src/" $ do
+            touch "src/ABC.hs"
+            touch "src/XYZ.hs"
+            [i|
+            library:
+              source-dirs: src
+              exposed-modules: Foo
+              generated-other-modules: ABC
+              generated-exposed-modules: XYZ
+            |] `shouldRenderTo` (library [i|
+            hs-source-dirs:
+                src
+            exposed-modules:
+                Foo
+                XYZ
+            other-modules:
+                Paths_foo
+                ABC
+            autogen-modules:
+                ABC
+                XYZ
+            |]) {packageCabalVersion = ">= 2.0"}
+
         context "with other-modules" $ do
           it "infers exposed-modules" $ do
             touch "src/Foo.hs"
@@ -498,6 +537,28 @@ spec = around_ (inTempDirectoryNamed "foo") $ do
             other-modules:
                 Bar
             |]
+
+          it "doesn't doubly include generated modules under src/" $ do
+            touch "src/ABC.hs"
+            touch "src/XYZ.hs"
+            [i|
+            library:
+              source-dirs: src
+              other-modules: Foo
+              generated-other-modules: ABC
+              generated-exposed-modules: XYZ
+            |] `shouldRenderTo` (library [i|
+            hs-source-dirs:
+                src
+            exposed-modules:
+                XYZ
+            other-modules:
+                Foo
+                ABC
+            autogen-modules:
+                ABC
+                XYZ
+            |]) {packageCabalVersion = ">= 2.0"}
 
         context "with both exposed-modules and other-modules" $ do
           it "doesn't infer any modules" $ do
@@ -533,6 +594,27 @@ spec = around_ (inTempDirectoryNamed "foo") $ do
             other-modules:
                 Paths_foo
             |]
+
+          it "doesn't doubly include generated modules under src/" $ do
+            touch "src/ABC.hs"
+            touch "src/XYZ.hs"
+            [i|
+            library:
+              source-dirs: src
+              generated-other-modules: ABC
+              generated-exposed-modules: XYZ
+            |] `shouldRenderTo` (library [i|
+            hs-source-dirs:
+                src
+            exposed-modules:
+                XYZ
+            other-modules:
+                Paths_foo
+                ABC
+            autogen-modules:
+                ABC
+                XYZ
+            |]) {packageCabalVersion = ">= 2.0"}
 
         context "with a conditional" $ do
           it "doesn't infer any modules mentioned in that conditional" $ do
@@ -600,6 +682,37 @@ spec = around_ (inTempDirectoryNamed "foo") $ do
                 hs-source-dirs:
                     unix/
               |]
+
+        it "includes all generated modules" $ do
+          [i|
+          library:
+            source-dirs: src
+            generated-exposed-modules: Exposed
+            generated-other-modules: Other
+            when:
+              condition: os(windows)
+              generated-exposed-modules: WinExposed
+              generated-other-modules: WinOther
+          |] `shouldRenderTo` (library [i|
+          hs-source-dirs:
+              src
+          if os(windows)
+            exposed-modules:
+                WinExposed
+            other-modules:
+                WinOther
+            autogen-modules:
+                WinOther
+                WinExposed
+          exposed-modules:
+              Exposed
+          other-modules:
+              Paths_foo
+              Other
+          autogen-modules:
+              Other
+              Exposed
+          |]) {packageCabalVersion = ">= 2.0"}
 
     context "with internal-libraries" $ do
       it "accepts internal-libraries" $ do
@@ -686,6 +799,26 @@ spec = around_ (inTempDirectoryNamed "foo") $ do
             other-modules:
                 Baz
           |]
+
+        it "doesn't doubly include generated modules under src/" $ do
+          touch "src/XYZ.hs"
+          [i|
+          executables:
+            foo:
+              main: Main.hs
+              source-dirs: src
+              generated-other-modules: XYZ
+          |] `shouldRenderTo` (executable "foo" [i|
+            main-is: Main.hs
+            hs-source-dirs:
+                src
+            other-modules:
+                Paths_foo
+                XYZ
+            autogen-modules:
+                XYZ
+          |]) {packageCabalVersion = ">= 2.0"}
+
         context "with conditional" $ do
           it "doesn't infer any modules mentioned in that conditional" $ do
             touch "src/Foo.hs"